chore(PT-51): move feature (booking) files to new folder structure
