+++ src/features/booking/components/BookingPrice.tsx
